Convert batch scripts to powershell, write validation error if zip invalid (#31)
See also: nvaccess/addon-datastore#1008

PowerShell has a bug with stripping quotes when passing to batch scripts (PowerShell/PowerShell#15250).
Additionally, no validation error was written to file if the downloaded file was not a valid zip file.

testing
See example nvaccess/addon-datastore-staging#11 where a comment was successfully posted and urls were successfully quoted via this action run

+++ validation/_validate/validate.py
@@ -362,6 +362,7 @@ def main():
 		help="The JSON file containing valid NVDA API versions."
 	)
 	parser.add_argument(
+		"--output",
 		dest="errorOutputFile",
 		help="The text file to output errors from the validation, if any.",
 		default=None,
@@ -371,7 +372,7 @@ def main():
 	addonFiles: List[str] = glob(args.filePathGlob)
 	verFilename: str = args.APIVersions
 	errorOutputFile: str = args.errorOutputFile
-	if os.path.exists(errorOutputFile):
+	if errorOutputFile and os.path.exists(errorOutputFile):
 		os.remove(errorOutputFile)
 
 	if not args.dry_run:
